% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.fodor2018.pass}
\alias{dat.fodor2018.pass}
\title{Meta-analytic data from Fodor et al. (2018): VR therapy vs. passive control groups}
\format{
An object of class \code{data.frame} with 26 rows and 5 columns.
}
\usage{
dat.fodor2018.pass
}
\description{
A data frame with 26 rows and 5 columns.
\itemize{
\item study: study identifier
\item ai: number of events (dropout) in the treatment group (VR therapy group)
\item ci: number of events (dropout) in the control group (passive control group)
\item n1i: sample size of the treatment group
\item n2i: sample size of the control group
}
}
\examples{
# load data frame:

data <- dat.fodor2018.pass

# calculate descriptives using rareDescribe():

descr <- rareDescribe(ai = ai,
                      n1i = n1i,
                      ci = ci,
                      n2i = n2i,
                      data = data)

print(descr)

# conduct random-effects meta-analysis
# using the inverse variance model:

ma <- rareIV(descr,
             measure = "logOR",
             method = "REML",
             cc = "constant",
             ccto = "only0",
             ccval = 0.5,
             drop00 = FALSE)

summary(ma)

}
\references{
Fodor, L. A., Coteț, C. D., Cuijpers, P., Szamoskozi, Ș., David, D., & Cristea, I. A. (2018).
The effectiveness of virtual reality based interventions for symptoms of anxiety and depression: A meta-analysis.
Scientific reports, 8(1), 10323. \href{https://doi.org/10.1038/s41598-018-28113-6}{SciRep}.
}
\keyword{datasets}
