% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.hoppen2023}
\alias{dat.hoppen2023}
\title{Meta-analytic data from Hoppen et al. (2023): EMDR vs. passive control conditions}
\format{
An object of class \code{data.frame} with 7 rows and 5 columns.
}
\usage{
dat.hoppen2023
}
\description{
A data frame with 7 rows and 5 columns.
\itemize{
\item study: study identifier
\item ai: number of events (dropout) in the treatment group (EMDR)
\item ci: number of events (dropout) in the control group (passive control conditions)
\item n1i: sample size of the treatment group
\item n2i: sample size of the control group
}
}
\examples{
# load data frame:

data <- dat.hoppen2023

# calculate descriptives using rareDescribe():

descr <- rareDescribe(ai = ai,
                      n1i = n1i,
                      ci = ci,
                      n2i = n2i,
                      data = data)

print(descr)

# conduct random-effects meta-analysis
# using the inverse variance model:

ma <- rareIV(descr,
             measure = "logOR",
             method = "REML",
             cc = "constant",
             ccto = "only0",
             ccval = 0.5,
             drop00 = FALSE)

summary(ma)

}
\references{
Hoppen, T. H., Jehn, M., Holling, H., Mutz, J., Kip, A., & Morina, N. (2023).
The efficacy and acceptability of psychological interventions for adult PTSD:
A network and pairwise meta-analysis of randomized controlled trials.
Journal of Consulting and Clinical Psychology.
\href{https://doi.org/10.1037/ccp0000809}{J CONSULT CLIN PSYCH}.
}
\keyword{datasets}
