% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.nissen2007}
\alias{dat.nissen2007}
\title{Meta-analytic data from Nissen and Wolski (2007)}
\format{
A data frame with 42 rows and 8 columns.
}
\usage{
dat.nissen2007
}
\description{
Meta-analytic data from Nissen and Wolski (2007)
}
\references{
Nissen and Wolski (2007) (\href{https://www.nejm.org/doi/full/10.1056/NEJMoa072761}{NEJM})
}
\keyword{datasets}
