% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.nissen2010}
\alias{dat.nissen2010}
\title{Meta-analytic data from Nissen and Wolski (2010)}
\format{
A data frame with 56 rows and 10 columns.
}
\usage{
dat.nissen2010
}
\description{
Meta-analytic data from Nissen and Wolski (2010)
}
\references{
Nissen and Wolski (2010) (\href{https://jamanetwork.com/journals/jamainternalmedicine/article-abstract/225844}{JAMA})
}
\keyword{datasets}
