% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.okumura2014}
\alias{dat.okumura2014}
\title{Meta-analytic data from Okumura et al. (2014): Group CBT for depression vs. non-active control}
\format{
An object of class \code{data.frame} with 27 rows and 5 columns.
}
\usage{
dat.okumura2014
}
\description{
A data frame with 27 rows and 5 columns.
\itemize{
\item study: study identifier
\item ai: number of events (dropout) in the treatment group (group CBT group)
\item ci: number of events (dropout) in the control group (non-active control group)
\item n1i: sample size of the treatment group
\item n2i: sample size of the control group
}
}
\examples{
# load data frame:

data <- dat.okumura2014

# calculate descriptives using rareDescribe():

descr <- rareDescribe(ai = ai,
                      n1i = n1i,
                      ci = ci,
                      n2i = n2i,
                      data = data)

print(descr)

# conduct random-effects meta-analysis
# using the inverse variance model:

ma <- rareIV(descr,
             measure = "logOR",
             method = "REML",
             cc = "constant",
             ccto = "only0",
             ccval = 0.5,
             drop00 = FALSE)

summary(ma)

}
\references{
Okumura, Y., & Ichikura, K. (2014). Efficacy and acceptability of
group cognitive behavioral therapy for depression: a systematic review and
meta-analysis. Journal of Affective Disorders, 164, 155-164.
\href{https://doi.org/10.1016/j.jad.2014.04.023}{JAD}.
}
\keyword{datasets}
