% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.yang2019}
\alias{dat.yang2019}
\title{Meta-analytic data from Yang et al. (2019): Psychological interventions for social anxiety disorder vs. control}
\format{
An object of class \code{data.frame} with 16 rows and 5 columns.
}
\usage{
dat.yang2019
}
\description{
A data frame with 16 rows and 5 columns.
\itemize{
\item study: study identifier
\item ai: number of events (dropout) in the treatment group (psychological intervention group)
\item ci: number of events (dropout) in the control group (control group)
\item n1i: sample size of the treatment group
\item n2i: sample size of the control group
}
}
\examples{
# load data frame:

data <- dat.yang2019

# calculate descriptives using rareDescribe():

descr <- rareDescribe(ai = ai,
                      n1i = n1i,
                      ci = ci,
                      n2i = n2i,
                      data = data)

print(descr)

# conduct random-effects meta-analysis
# using the inverse variance model:

ma <- rareIV(descr,
             measure = "logOR",
             method = "REML",
             cc = "constant",
             ccto = "only0",
             ccval = 0.5,
             drop00 = FALSE)

summary(ma)

}
\references{
Yang, L., Zhou, X., Pu, J., Liu, L., Cuijpers, P., Zhang, Y., ... & Xie, P. (2019).
Efficacy and acceptability of psychological interventions for social anxiety disorder in
children and adolescents: a meta-analysis of randomized controlled trials.
European Child & Adolescent Psychiatry, 28, 79-89.
\href{https://doi.org/10.1007/s00787-018-1189-x}{EUR CHILD ADOLES PSY}.
}
\keyword{datasets}
