% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.yuan2018}
\alias{dat.yuan2018}
\title{Meta-analytic data from Yuan et al. (2018)}
\format{
An object of class \code{data.frame} with 8 rows and 5 columns.
}
\usage{
dat.yuan2018
}
\description{
A data frame with 8 rows and 5 columns.
\itemize{
\item study: study identifier
\item ai: number of events (dropout) in the treatment group
\item ci: number of events (dropout) in the control group
\item n1i: sample size of the treatment group
\item n2i: sample size of the control group
}
}
\examples{
# load data frame:

data <- dat.yuan2018

# calculate descriptives using rareDescribe():

descr <- rareDescribe(ai = ai,
                      n1i = n1i,
                      ci = ci,
                      n2i = n2i,
                      data = data)

print(descr)

# conduct random-effects meta-analysis
# using the inverse variance model:

ma <- rareIV(descr,
             measure = "logOR",
             method = "REML",
             cc = "constant",
             ccto = "only0",
             ccval = 0.5,
             drop00 = FALSE)

summary(ma)

}
\references{
Yuan, S., Zhou, X., Zhang, Y., Zhang, H., Pu, J., Yang, L., Liu, L., Jiang, X., & Xie, P. (2018). Comparative efficacy
and acceptability of bibliotherapy for depression and anxiety disorders in children and adolescents: A meta-analysis
of randomized clinical trials. Neuropsychiatric Disease and Treatment, 14, 353–365. \href{https://doi.org/10.2147/NDT.S152747}{NDT}.
}
\keyword{datasets}
