% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{equalLength}
\alias{equalLength}
\title{Check if lengths of vectors are equal}
\usage{
equalLength(...)
}
\arguments{
\item{...}{A collection of vectors, separated by a comma.}
}
\value{
A logical that indicates whether the lengths of the vectors are equal.
}
\description{
Check if lengths of vectors are equal
}
