% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareData.R
\name{new_rareData}
\alias{new_rareData}
\title{Constructor function for class rareData}
\usage{
new_rareData(x = list())
}
\arguments{
\item{x}{a list containing the elements required for objects of the class rareData}
}
\value{
an object of class rareData
}
\description{
Constructor function for class rareData
}
