% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raremeta.R
\name{new_raremeta}
\alias{new_raremeta}
\title{Constructor function for class raremeta}
\usage{
new_raremeta(x = list())
}
\arguments{
\item{x}{a list containing the elements required for objects of the class raremeta}
}
\value{
an object of class raremeta
}
\description{
Constructor function for class raremeta
}
