% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rareData.R
\name{print.rareData}
\alias{print.rareData}
\title{Print Method for rareData objects.}
\usage{
\method{print}{rareData}(x, ...)
}
\arguments{
\item{x}{an object of class rareData}

\item{...}{other arguments.}
}
\value{
The print functions do not return an object
}
\description{
Print method for objects of class "rareData".
}
