% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.raremeta.R
\name{print.raremeta}
\alias{print.raremeta}
\title{Print Method for raremeta objects.}
\usage{
\method{print}{raremeta}(x, digits, ...)
}
\arguments{
\item{x}{an object of class raremeta}

\item{digits}{number of digits to be shown in the print output}

\item{...}{other arguments.}
}
\value{
The print functions do not return an object
}
\description{
Print method for objects of class "raremeta".
}
