% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareBetabin.R
\name{rareBetabin}
\alias{rareBetabin}
\title{Conduct a meta-analysis using the beta-binomial model}
\usage{
rareBetabin(
  x,
  measure,
  common_rho = TRUE,
  drop00 = FALSE,
  level = 95,
  test = "z",
  digits = 4,
  verbose = FALSE,
  control,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{measure}{character string specifying the effect size or outcome measure to be used
(either \code{"logOR"} for the log odds ratio or \code{"logRR"} for the log relative risk).}

\item{common_rho}{logical specifying whether a common intraclass correlations shall be assumed
for the two groups (\code{TRUE}), or whether the intraclass correlation shall be allowed
to differ between the two groups (\code{FALSE})}

\item{drop00}{logical indicating whether double-zero studies (i.e., studies with no events or
only events in both groups) should be excluded when calculating the outcome measufit for the
individual studies.}

\item{level}{numeric between 0 and 100 specifying the confidence interval level (the default is 95).}

\item{test}{character string specifying how test statistics and confidence intervals for the
fixed effects should be computed (currently, only \code{"z"}, for Wald-type tests is available).}

\item{digits}{integer specifying the number of decimal places to which the printed results
should be rounded (if unspecified, the default is 4).}

\item{verbose}{logical indicating whether output should be generated on the progress of model
fitting (the default is \code{FALSE}). Can also be an integer. Values > 1 generate more verbose output.}

\item{control}{optional list of control values for the iterative algorithms. If unspecified, default
values are defined inside the functions.}

\item{...}{additional arguments.}
}
\value{
an object of class "raremeta". The object is a list containing the following elements:
\itemize{
\item \code{model}: name of the model used for conducting the meta-analysis.
\item \code{beta}: estimated coefficients of the model.
\item \code{se}: standard errors of the  coefficients.
\item \code{zval}: test statistics of the coefficients.
\item \code{zval}: p-values corresponding to the test statistics.
\item \code{ci.lb}: lower bound of the confidence intervals for the coefficients.
\item \code{ci.ub}: upper bound of the confidence intervals for the coefficients.
\item \code{vb}: variance-covariance matrix of the estimated coefficients.
\item \code{rho}: estimated intraclass correlation. If \code{common_rho = FALSE}, rho is a vector
which contains both intraclass correlations.
\item \code{LRT.Chisq}: Test statistic of the likelihood ratio test testing for homogeneity.
\item \code{LRT.df}: Degrees of freedom of the likelihood ratio test testing for homogeneity.
\item \code{LRT.pval}: p-value of the likelihood ratio test testing for homogeneity.
\item \code{fit.stats}: a list with log-likelihood, deviance, AIC, BIC, and AICc values under
the unrestricted and restricted likelihood.
\item \code{p}: number of coefficients in the model (including the intercept).
\item \code{k}: number of studies included in the analysis.
\item \code{k.all}: total number of studies (before exclusion).
\item \code{kdz},\code{ksz}: number of double-zero and single-zero studies.
\item \code{k1sz}, \code{k2sz}: number of single-zero studies where the zero is in group 1 or group 2.
\item \code{ai}, \code{bi}, \code{ci}, \code{di}: original entries of the 2x2 tables for all studies.
\item \code{ni}, \code{n1i}, \code{n2i}: original total and group sample sizes.
\item ...
}
}
\description{
Function to conduct a meta-analysis of a rare event using a beta-binomial model
See below for more details on this model and its application in meta-analyses of
rare events.
}
\section{Details}{
...
}

