% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareCC.R
\name{rareCC}
\alias{rareCC}
\title{Apply continuity correction to your data}
\usage{
rareCC(
  x,
  ai,
  bi,
  ci,
  di,
  n1i,
  n2i,
  data,
  cc = "constant",
  ccval = 0.5,
  tccval,
  cccval,
  ccsum = 1,
  ccto = "only0",
  drop00 = TRUE,
  measure,
  method = "FE"
)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{ai}{Data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{Data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{Data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{Data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{Data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{Data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{Data frame.}

\item{cc}{character string specifying the type of continuity corrections to be used
(either \code{"constant"}, \code{"tacc"} or \code{"empirical"}). Default is "constant". See 'Details'.}

\item{ccval}{scalar or numerical vector specifying the value of the continuity correction if
\code{cc = "constant"}. Must be a scalar or a vector of length equal to the number of studies.
Default is \code{ccval = 0.5}. If a scalar is specified, the value is added to all studies for
which the number of events is zero in at least one of the groups. This behavior can be changed
using the argument \code{ccto}. The argument \code{ccval} is overwritten by \code{tccval} and \code{cccval} if both arguments are specified.}

\item{tccval}{scalar or numerical vector specifying the value of the continuity correction
applied to the observations from the treatment group (group 1) if \code{cc = "constant"}. Must be a scalar or a vector
of length equal to the number of studies. If \code{cc = "constant"} and \code{tccval} is not specified, \code{tccval} is
set to the value of \code{ccval} internally.}

\item{cccval}{scalar or numerical vector specifying the value of the continuity correction
applied to the observations from the control group (group 2) if \code{cc = "constant"}. Must be a scalar or a vector
of length equal to the number of studies. If \code{cc = "constant"} and \code{cccval} is not specified, \code{cccval} is
set to the value of \code{ccval} internally.}

\item{ccsum}{numeric value specifying the value of the sum of the continuity correction applied to the
observations from the treatment group and the continuity correction applied to the observations from
the control group. Default is \code{ccsum = 1}. Currently, setting this argument to a different number only has
an effect when \code{cc = "tacc"} or \code{cc = "empirical"}.}

\item{ccto}{character string indicating to which studies the continuity correction should
be applied. Either \code{"only0"}, for which the continuity correction is applied to all studies
for which the number of events is zero in at least one of the groups, \code{"all"}, for which the
continuity correction is applied to all studies, or \code{"if0all"}, for which the continuity
correction is applied to all studies if any of the individual studies has zero events in at
least one of the groups.}

\item{drop00}{logical indicating whether double-zero studies (i.e., studies with no events or
only events in both groups) should be excluded from the dataset.}

\item{measure}{character string specifying the effect size or outcome measure to be used
(either \code{"logOR"} for the log odds ratio, \code{"logRR"} for the log relative risk,
or \code{"RD"} for the risk difference). Important when selecting continuity corrections dependent
on the estimated summary effect size of some subset of the studies involved in the analyisis,
i.e. \verb{cc = "empircal}.}

\item{method}{character string specifying whether a fixed- or a random-effects model should be fitted.
A fixed-effects model is fitted when using \code{method = "FE"} . A random-effects model is fitted
by setting \code{method} equal to one of the following: \code{"DL"}, \code{"HE"}, \code{"SJ"}, \code{"ML"}, \code{"REML"}, \code{"EB"}, \code{"HS"},
\code{"PM"}, \code{"IPM"}, \code{"GENQ"}, \code{"PMM"} or \code{"GENQM"}. Important when selecting continuity corrections dependent
on the estimated summary effect size of some subset of the studies involved in the analyisis,
i.e. \code{cc = "empircal"}.}
}
\value{
an object of class "raremeta". The object is a list containing the following elements:
\itemize{
\item \code{ai}, \code{bi}, \code{ci}, \code{di}: original entries of the 2x2 tables for all studies.
\item \code{ai.cc}, \code{bi.cc}, \code{ci.cc}, \code{di.cc}: entries of the 2x2 tables for all studies after
application of the specified continuity correction.
\item \code{ni}, \code{n1i}, \code{n2i}: original total and group sample sizes.
\item \code{ni.cc}, \code{n1i.cc}, \code{n2i.cc}: total and group sample sizes after application of the
specified continuity correction.
\item \code{tcc}, \code{ccc}: value of the specified continuity correction for the treatment group (group 1) and control
group (group 2).
\item \code{cc.studies}: vector which indicates whether the continuity correction was applied
to a study.
\item \code{remove}: logical vector indicating which studies were removed before
aplication of the continuity correction
\item \code{kdz},\code{ksz}: number of double-zero and single-zero studies.
\item \code{k1sz}, \code{k2sz}: number of single-zero studies where the zero is in group 1 or group 2.
\item \code{cc} : the type of continuity correction applied
\item \code{drop00}: logical indicating whether double-zero studies were omitted
\item ...
}
}
\description{
Function to apply different kinds of continuity correction to your data.
}
\section{Details}{
\subsection{Data input}{

Data input can happen either through the parameter \code{x} (an object of type \code{rareData})
or through the parameters \code{ai},\code{bi},\code{ci}, \code{di}, \code{n1i}, \code{n2i}, \code{data} (columns of a dataframe).
A \code{rareData} object can be produced from a data frame by applying the \code{rareDescribe()} function to it. The \code{rareDescribe()}
function pre-processes the data frame and stores the information required by the \code{rareCC()} function
in a list. See \code{?rareDescribe} for more details.
}

\subsection{Types of continuity correction}{

This function offers three kinds of continuity correction.
\subsection{Constant continuity correction}{

When setting \code{cc = "constant"}, a constant value will be added to all
cells of all studies specified via the \code{ccto} argument. The default is \code{ccval = 0.5}.
Through specification of \code{tccval} and \code{cccval} different values can be specified for the
cells of the treatment group and the control group, respectively.
By inputting vectors in the afforementioned arguments, different values can be added
to different studies (adding the \code{k}-th value to the \code{k}-th study),
making it possible for the user to introduce their own continuity correction.
}

\subsection{Treatment-arm continuity correction}{

When setting \code{cc = "tacc"} (treatment-arm continuity correction), the size of
both of study arms is used to calculate a value for continuity correction.
For a precise definition see Sweeting et al. (2004).
}

\subsection{Empirical correction}{

When setting \code{cc = "empirical"}, the estimated summary effect size
(logOR, logRR or RD) for all studies which enable the estimation of the corresponding individual effect
size is used to calculate a value for continuity correction.
This means that there must be at least one study enabling the estimation of
the corresponding individual effect size and the \code{measure}- and \code{method} argument must be
specified.
For a precise definition see Sweeting et al. (2004).
When it comes to model fitting, there is the possibility to fit fixed-effects models (also known as equal-effects models)
and random-effects models using the inverse variance approach.
A fixed-effects model is fitted when \code{method} is set to \code{"FE"} (or \code{"EE"}). A random-effects model
is fitted when \code{method} is set to either \code{"DL"}, \code{"HE"}, \code{"SJ"}, \code{"ML"}, \code{"REML"}, \code{"EB"}, \code{"HS"},
\code{"PM"}, \code{"IPM"}, \code{"GENQ"}, \code{"PMM"} or \code{"GENQM"}.
For a basic introduction to fixed-effects and random-effects meta-analysis, please refer to Borenstein et al. (2021)
Currently, the model fit is done by using the \code{metafor} package, see Viechtbauer (2010).
}

}

\subsection{Estimation of the between-study variance in random-effects meta-analysis}{

Different estimators can be used to estimate the between-study variance, tau^2, in random-effects
meta-analysis. The estimator to be used is specified via the \code{methods} argument.
\itemize{
\item \code{"DL"}: DerSimonian-Laird estimator
\item \code{"HE"}: Hedges estimator
\item \code{"SJ"}: Sidik-Jonkman estimator
\item \code{"ML"}: maximum likelihood estimator
\item \code{"REML"}: restricted maximum likelihood estimator
\item \code{"EB"}: empirical Bayes estimator
\item \code{"HS"}: Hunter-Schmidt estimator
\item \code{"PM"}: Paule-Mandel estimator
\item \code{"IPM"}: improved Paule-Mandel estimator
\item \code{"GENQ"}: generalized Q-statistic estimator
\item \code{"PMM"}: median-unbiased Paule-Mandel estimator
\item \code{"GENQM"}: median-unbiased generalized Q-statistic estimator
}

Most of these estimators are described in Zhang et al. (2021). For details on the improved Paule-Mandel estimator,
see also Bhaumik et al. (2012). The median-unbiased Paule-Mandel estimator and the median-unbiased generalized
Q-statistic estimator are described in Viechtbauer (2021).
}
}

\examples{
#initializing a data set
data("dat.nissen2007")
d <- dat.nissen2007
x <- rareDescribe(ai=miRosiglitazone, n1i=nRosiglitazone, ci=miControl, n2i=nControl, data= d)

#applying the default continuity correction (i.e. cc = "constant" with ccval = 0.5)
x.constant <- rareCC(x)

#applying the treatment-arm continuity correction with `ccsum = 0.1`
x.tacc <- rareCC(x, cc = "tacc", ccsum = 0.1)

#applying the empirical continuity correction for the log odds ratio for the fixed effects model
x.empirical <- rareCC(x, cc = "empirical", measure = "logOR", method = "FE")

}
\references{
Borenstein, M., Hedges, L. V., Higgins, J. P., & Rothstein, H. R. (2021). Introduction to meta-analysis.
John Wiley & Sons.

Bhaumik, D. K., Amatya, A., Normand, S.-L. T., Greenhouse, J., Kaizar, E., Neelon, B., & Gibbons, R. D. (2012).
Meta-analysis of rare binary adverse event data. Journal of the American Statistical
Association 107, 555–567. doi: 10.1080/01621459.2012.664484

Sweeting, M. J., Sutton, A. J., & Lambert, P. C. (2004). What to add to nothing? Use and avoidance of
continuity corrections in meta-analysis of sparse data. Statistics in Medicine, 23, 1351–1375. doi: 10.1002/sim.1761

Viechtbauer, W. (2010). Conducting Meta-Analyses in R with the metafor Package.
Journal of Statistical Software, 36(3), 1–48. https://doi.org/10.18637/jss.v036.i03

Viechtbauer, W. (2021). Median-unbiased estimators for the amount of heterogeneity in meta-analysis. European Congress of Methodology,
Valencia, Spain. https://www.wvbauer.com/lib/exe/fetch.php/talks:2021_viechtbauer_eam_median_tau2.pdf

Zhang, C., Chen, M., & Wang, X. (2020). Statistical methods for quantifying between-study heterogeneity in meta-analysis
with focus on rare binary events. Statistics and its interface, 13(4), 449. doi: 10.4310/sii.2020.v13.n4.a3
}
