% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareData.R
\name{rareData}
\alias{rareData}
\title{Helper function to generate objects of class rareData}
\usage{
rareData(x = list())
}
\arguments{
\item{x}{a list containing the elements required for objects of the class rareData}
}
\value{
an object of class rareData
}
\description{
Helper function to generate objects of class rareData
}
