% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareHet.R
\name{rareHet}
\alias{rareHet}
\title{Testing for between-study homogeneity in meta-analyses of rare events}
\usage{
rareHet(
  x,
  ai,
  bi,
  ci,
  di,
  n1i,
  n2i,
  data,
  measure = "logOR",
  method = "FE",
  test = "q",
  type = "standard",
  approx = FALSE,
  cc,
  ccval = 0.5,
  tccval,
  cccval,
  ccsum = 1,
  ccto = "only0",
  drop00 = TRUE
)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{ai}{data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{data frame}

\item{measure}{character string specifying the effect size or outcome measure to be used
(either \code{"logOR"} for the log odds ratio, \code{"logRR"} for the log relative risk,
or \code{"RD"} for the risk difference).}

\item{method}{character string specifying whether the test shall be conducted in  a
fixed- or a random-effects framework. The test is based on a fixed-effects model when specifying \code{method = "FE"} .
The test is carried out in a random-effects framework if \code{method} equal to one of the following: \code{"DL"}, \code{"HE"},
\code{"SJ"}, \code{"ML"}, \code{"REML"}, \code{"EB"}, \code{"HS"}, \code{"PM"}, \code{"IPM"}, \code{"GENQ"}, \code{"PMM"} or \code{"GENQM"}.}

\item{test}{type of test for between-study homogeneity, either \code{"q"} (for a Q test),
\code{"lrt"} (for a likelihood-ratio test), \code{"score"} (for a score test), \code{"wald"} (for a Wald test).
Default is \code{q}.}

\item{type}{specific type of test to be used. For \code{test = "q"},
\code{type} can be set to \code{"standard"}, \code{"modified"}, \verb{"jackson1", "jackson2"}, \code{"gart"}, \code{"bliss"}, or
\code{"kd"} (short for Kulinskaya-Dollinger). For \code{test = "lrt"}, \code{test = "score"} or \code{test = "wald"},
\code{method} can be set to \code{"ML"} or \code{"REML"}. See Details.}

\item{approx}{logical to specify whether an approximate distribution shall be used if possible}

\item{cc}{character string specifying the type of continuity corrections to be used
(either \code{"constant"}, \code{"tacc"} or \code{"empirical"}). Default is "constant".}

\item{ccval}{scalar or numerical vector specifying the value of the continuity correction if
\code{cc = "constant"}. Must be a scalar or a vector of length equal to the number of studies.
Default is \code{ccval = 0.5}. If a scalar is specified, the value is added to all studies for
which the number of events is zero in at least one of the groups. This behavior can be changed
using the argument \code{ccto}. \code{ccval} is overwritten by tccval and cccval if both arguments are specified.}

\item{tccval}{scalar or numerical vector specifying the value of the continuity correction
applied to the observations from the treatment group (group 1) if \code{cc = "constant"}. Must be a scalar or a vector
of length equal to the number of studies. If \code{cc = "constant"} and \code{tccval} is not specified, \code{tccval} is
set to the value of \code{ccval} internally.}

\item{cccval}{scalar or numerical vector specifying the value of the continuity correction
applied to the observations from the control group (group 2) if \code{cc = "constant"}. Must be a scalar or a vector
of length equal to the number of studies. If \code{cc = "constant"} and \code{cccval} is not specified, \code{cccval} is
set to the value of \code{ccval} internally.}

\item{ccsum}{numeric value specifying the value of the sum of the continuity correction applied to the
observations from the treatment group and the continuity correction applied to the observations from
the control group. Default is \code{ccsum = 1}. Currently, setting this argument to a different number only has
an effect when \code{cc = "tacc"} or \code{cc = "empirical"}.}

\item{ccto}{character string indicating to which studies the continuity correction should
be applied. Either \code{"only0"}, for which the continuity correction is applied to all studies
for which the number of events is zero in at least one of the groups, \code{"all"}, for which the
continuity correction is applied to all studies, or \code{"if0all"}, for which the continuity
correction is applied to all studies if any of the individual studies has zero events in at
least one of the groups.}

\item{drop00}{logical indicating whether double-zero studies (i.e., studies with no events or
only events in both groups) should be excluded when calculating the outcome measufit for the
individual studies.
(this option is available for \code{type = "score"} and \code{method = "cond"} or \code{method = "BD"}).}
}
\value{
a list including the value of the test statistic and the associated \strong{p} value.
\subsection{Data input}{

Data input can happen either through the parameter \code{x} (an object of type \code{rareData})
or through the parameters \code{ai},\code{bi},\code{ci}, \code{di}, \code{n1i}, \code{n2i}, \code{data} (columns of a dataframe).
A \code{rareData} object can be produced from a data frame by applying the \code{rareDescribe()} function to it.
The \code{rareDescribe()} function pre-processes the data frame and stores the information required by the \code{rareES()} function
in a list. See \code{?rareDescribe} for more details.
}
}
\description{
Testing for between-study homogeneity in meta-analyses of rare events
}
\references{
Zhang, C., Chen, M., & Wang, X. (2020). Statistical methods for quantifying between-study heterogeneity in meta-analysis
with focus on rare binary events. Statistics and its interface, 13(4), 449. doi: 10.4310/sii.2020.v13.n4.a3
}
