% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareIV.R
\name{rareIV}
\alias{rareIV}
\title{Conduct a meta-analysis of a rare event using the inverse variance model}
\usage{
rareIV(
  x,
  measure,
  method,
  cc,
  ccval = 0.5,
  tccval,
  cccval,
  ccsum = 1,
  ccto = "only0",
  drop00 = TRUE,
  weighted = TRUE,
  level = 95,
  test = "z",
  digits = 4,
  verbose = FALSE,
  control,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{measure}{character string specifying the effect size or outcome measure to be used
(either \code{"logOR"} for the log odds ratio, \code{"logRR"} for the log relative risk,
or \code{"RD"} for the risk difference).}

\item{method}{character string specifying whether a fixed- or a random-effects model should be fitted.
A fixed-effects model is fitted when using \code{method = "FE"}. A random-effects model is fitted
by setting \code{method} equal to one of the following: "DL", "HE", "SJ", "ML", "REML", "EB", "HS", "PM", "IPM", or "GENQ".
Default is \code{"SJ"}.}

\item{cc}{character string specifying the type of continuity corrections to be used
(either \code{"constant"}, \code{"tacc"} or \code{"empirical"}). Default is "constant". See 'Details'.}

\item{ccval}{scalar or numerical vector specifying the value of the continuity correction if
\code{cc = "constant"}. Must be a scalar or a vector of length equal to the number of studies.
Default is \code{ccval = 0.5}. If a scalar is specified, the value is added to all studies for
which the number of events is zero in at least one of the groups. This behavior can be changed
using the argument \code{ccto}. \code{ccval} is overwritten by tccval and cccval if both arguments are specified.}

\item{tccval}{scalar or numerical vector specifying the value of the continuity correction
applied to the observations from the treatment group if \code{cc = "constant"}. Must be a scalar or a vector
of length equal to the number of studies. If \code{cc = "constant"} and \code{tccval} is not specified, \code{tccval} is
set to the value of \code{ccval} internally.}

\item{cccval}{scalar or numerical vector specifying the value of the continuity correction
applied to the observations from the control group if \code{cc = "constant"}. Must be a scalar or a vector
of length equal to the number of studies. If \code{cc = "constant"} and \code{cccval} is not specified, \code{cccval} is
set to the value of \code{ccval} internally.}

\item{ccsum}{numeric value specifying the value of the sum of the continuity correction applied to the
observations from the treatment group and the continuity correction applied to the observations from
the control group. Default is \code{ccsum = 1}. Currently, setting this argument to a different number only has
an effect when \code{cc = "tacc"} or \code{cc = "empirical"}.}

\item{ccto}{character string indicating to which studies the continuity correction should
be applied. Either \code{"only0"}, for which the continuity correction is applied to all studies
for which the number of events is zero in at least one of the groups, \code{"all"}, for which the
continuity correction is applied to all studies, or \code{"if0all"}, for which the continuity
correction is applied to all studies if any of the individual studies has zero events in at
least one of the groups.}

\item{drop00}{logical indicating whether double-zero studies (i.e., studies with no events or
only events in both groups) should be excluded when calculating the outcome measufit for the
individual studies.}

\item{weighted}{logical specifying whether weighted (default) or unweighted estimation of the pooled
effect size should be used.}

\item{level}{numeric between 0 and 95 specifying the confdeince interval level (the default is 95)}

\item{test}{character string specifying how test statistics and confidence intervals for the
fixed effects should be computed (either \code{"z"}, for Wald-type tests and CIs, or \code{"hksj"}, for
tests and CIs based on the method by Knapp and Hartung (2003) and Sidik and Jonkman (2002).
Specifying \code{test = "knha"} instead of \code{test = "hksj"} will produce the same results).}

\item{digits}{integer specifying the number of decimal places to which the printed results
should be rounded (if unspecified, the default is 4).}

\item{verbose}{logical indicating whether output should be generated on the progress of model
fitting (the default is \code{FALSE}). Can also be an integer. Values > 1 generate more verbose output.}

\item{control}{optional list of control values for the iterative algorithms. If unspecified, default
values are defined inside the functions.}

\item{...}{additional arguments.}
}
\value{
an object of class raremeta. The object is a list containing the following elements:

...
}
\description{
Function to conduct a meta-analysis of a rare event using the fixed- or
random-effects model of the inverse variance approach.
}
\section{Details}{
\subsection{Data input}{
}

\subsection{Effect size measures}{
}

\subsection{Handling single-zero and double-zero studies}{
}

\subsection{Differences between effect size measures in the application of continuity corrections}{
}

\subsection{Model specification}{
}

\subsection{Estimation of the between-study variance in random-effects meta-analysis}{
}

\subsection{Tests and confidence intervals for model coefficients}{
}

\subsection{Tests for residual heterogeneity}{
}
}

\references{
Hartung, J., and Knapp, G. (2001). A refined method for the meta-analysis of controlled clinical trials
with binary outcome. Statistics in Medicine 20, 3875-3889. doi: 10.1002/sim.1009

Sidik, K., and Jonkman, J. N. (2002). A simple confidence interval for meta-analysis.
Statistics in Medicine 21, 3153-3159. doi: 10.1002/sim.1262
}
