% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raremeta.R
\name{raremeta}
\alias{raremeta}
\title{Helper function to generate objects of class raremeta}
\usage{
raremeta(x = list())
}
\arguments{
\item{x}{a list containing the elements required for objects of the class raremeta}
}
\value{
an object of class raremeta
}
\description{
Helper function to generate objects of class raremeta
}
