% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarePeto.R
\name{rarePeto}
\alias{rarePeto}
\title{Conduct a meta-analysis using Peto's method}
\usage{
rarePeto(x, ai, bi, ci, di, n1i, n2i, data, level = 95, digits = 4)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{ai}{data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{data frame.}

\item{level}{numeric nbetween 0 and 100 specifying the confidence interval
level (the default is 95)}

\item{digits}{integer specifying the number of decimal places to which the printed results
should be rounded (if unspecified, the default is 4).}
}
\value{
an object of class "raremeta".
The object is a list containing the following elements:
\itemize{
\item \code{beta}, \code{b}: estimated effect size.
\item \code{se}: standard error of the  estimator.
\item \code{zval}: test statistics of the coefficients.
\item \code{pval}: p-values corresponding to the test statistics.
\item \code{ci.lb}: lower bound of the confidence intervals for the coefficients.
\item \code{ci.ub}: upper bound of the confidence intervals for the coefficients.
\item \code{k}: number of studies included in the analysis.
\item \code{kdz},\code{ksz}: number of double-zero and single-zero studies.
\item \code{k1sz}, \code{k2sz}: number of single-zero studies where the zero is in group 1 or group 2.
\item \code{ai}, \code{bi}, \code{ci}, \code{di}: original entries of the 2x2 tables for all studies.
\item \code{ni}, \code{n1i}, \code{n2i}: original total and group sample sizes.
\item ...
}
}
\description{
Function to conduct a meta-analysis.
Effect size is the log odds ratio estimated from event counts in form of
2x2 contingency tables using Peto's method (see Yusuf et al., 1985).
}
\section{Details}{
\subsection{Data input}{

Data input can happen either through the parameters \code{ai},\code{bi},\code{ci},\code{di},\code{n1i},\code{n2i} (columns of the data frame \code{data})
or pre-processed throuth the parameter \code{x} (an object of type \code{rareData}).
A \code{rareData} object can be produced from a data frame by applying the \code{rareDescribe()} function to it.
The \code{rareDescribe()} function pre-processes the data frame and stores the information required by the \code{rarePeto()} function in a list.
See \code{?rareDescribe} for more details.
}

\subsection{Effect size measure}{

The function is a meta-analytic method for estimating log odds ratios.
Data comes in form of event counts in form of 2x2 contingency tables.\tabular{lrr}{
    \tab event \tab no event \cr
   group1 (treatment) \tab ai \tab bi \cr
   group2 (control) \tab ci \tab di \cr
}

}
}

\examples{
data <- data.frame(
ai = c(0, 3, 2, 0),
bi = c(20, 18, 15, 19),
ci = c(1, 4, 0, 0),
di = c(19, 17, 16, 20)
)

x <- rareDescribe(ai = ai, bi = bi, ci = ci, di = di, data = data)

rarePeto(x)

}
\references{
Yusuf, S., Peto, R., Lewis, J., Collins, R., & Sleight, P. (1985).
Beta blockade during and after myocardial infarction: an overview of
the randomized trials.
Progress in cardiovascular diseases, 27(5), 335-371.
}
