% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.raremeta.R
\name{summary.raremeta}
\alias{summary.raremeta}
\title{Summary Method for raremeta objects.}
\usage{
\method{summary}{raremeta}(object, ...)
}
\arguments{
\item{object}{an object of class raremeta}

\item{...}{other arguments.}
}
\value{
An object of class raremeta
}
\description{
Summary method for objects of class "raremeta".
}
