% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareData.R
\name{validate_rareData}
\alias{validate_rareData}
\title{validator function for class rareData}
\usage{
validate_rareData(x)
}
\arguments{
\item{x}{an object of class rareData}
}
\value{
an object of class rareData
}
\description{
validator function for class rareData
}
