% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raremeta.R
\name{validate_raremeta}
\alias{validate_raremeta}
\title{validator function for class raremeta}
\usage{
validate_raremeta(x)
}
\arguments{
\item{x}{an object of class raremeta}
}
\value{
an object of class raremeta
}
\description{
validator function for class raremeta
}
