% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.nissen2007}
\alias{dat.nissen2007}
\title{Meta-analytic data from Nissen and Wolski (2007)}
\format{
An object of class \code{data.frame} with 42 rows and 8 columns.
}
\usage{
dat.nissen2007

# load data frame:

data <- dat.nissen2007

# calculate descriptives using rareDescribe():

descr <- rareDescribe(ai = miRosiglitazone,
                      n1i = nRosiglitazone,
                      ci = miControl,
                      n2i = nControl,
                      data = data)

print(descr)

# conduct random-effects meta-analysis
# using the inverse variance model:

ma <- rareIV(descr,
             measure = "logOR",
             method = "REML",
             cc = "constant",
             ccto = "only0",
             ccval = 0.5,
             drop00 = FALSE)

summary(ma)
}
\description{
A data frame with 42 rows and 8 columns.
\itemize{
\item study: study identifier
\item nRosiglitazone: number of patients in the Rosiglitazone group
\item miRosiglitazone: number of myocardial infarctions in the Rosiglitazone group
\item cvRosiglitazone: number of deaths from cardiovascular causes in the Rosiglitazone group
\item nControl: number of patients in the control group
\item miControl: number of myocardial infarctions in the control group
\item cvControl: number of deaths from cardiovascular causes in the control group
\item methodControl: type of control group
}
}
\references{
Nissen and Wolski (2007) (\href{https://www.nejm.org/doi/full/10.1056/NEJMoa072761}{NEJM})
}
\keyword{datasets}
