% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareBetabin.R
\name{rareBetabin}
\alias{rareBetabin}
\title{Conduct a meta-analysis using the beta-binomial model}
\usage{
rareBetabin(
  x,
  ai,
  bi,
  ci,
  di,
  n1i,
  n2i,
  data,
  measure,
  common_rho = TRUE,
  drop00 = FALSE,
  level = 95,
  test = "z",
  digits = 4,
  verbose = FALSE,
  control,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{ai}{data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{data frame.}

\item{measure}{character string specifying the effect size or outcome measure to be used
(either \code{"logOR"} for the log odds ratio or \code{"logRR"} for the log relative risk).}

\item{common_rho}{logical specifying whether a common intraclass correlations shall be assumed
for the two groups (\code{TRUE}), or whether the intraclass correlation shall be allowed
to differ between the two groups (\code{FALSE}). See below for more detail.s}

\item{drop00}{logical indicating whether double-zero studies (i.e., studies with no events or only events in both groups)
should be excluded prior to calculating the studies' effect sizes and sampling variances.}

\item{level}{numeric between 0 and 100 specifying the confidence interval level (the default is 95).}

\item{test}{character string specifying how test statistics and confidence intervals for the
fixed effects should be computed (currently, only \code{"z"}, for Wald-type tests is available).}

\item{digits}{integer specifying the number of decimal places to which the printed results
should be rounded (if unspecified, the default is 4).}

\item{verbose}{logical indicating whether output should be generated on the progress of model
fitting (the default is \code{FALSE}). Can also be an integer. Values > 1 generate more verbose output.}

\item{control}{optional list of control values for the iterative algorithms. If unspecified, default
values are defined inside the functions.}

\item{...}{additional arguments.}
}
\value{
an object of class "raremeta". The object is a list containing the following elements:
\itemize{
\item \code{model}: name of the model used for conducting the meta-analysis.
\item \code{beta}: estimated coefficients of the model.
\item \code{se}: standard errors of the  coefficients.
\item \code{zval}: test statistics of the coefficients.
\item \code{zval}: p-values corresponding to the test statistics.
\item \code{ci.lb}: lower bound of the confidence intervals for the coefficients.
\item \code{ci.ub}: upper bound of the confidence intervals for the coefficients.
\item \code{vb}: variance-covariance matrix of the estimated coefficients.
\item \code{rho}: estimated intraclass correlation. If \code{common_rho = FALSE}, rho is a vector
which contains both intraclass correlations.
\item \code{LRT.Chisq}: Test statistic of the likelihood ratio test testing for homogeneity.
\item \code{LRT.df}: Degrees of freedom of the likelihood ratio test testing for homogeneity.
\item \code{LRT.pval}: p-value of the likelihood ratio test testing for homogeneity.
\item \code{fit.stats}: a list with log-likelihood, deviance, AIC, BIC, and AICc values under
the unrestricted and restricted likelihood.
\item \code{p}: number of coefficients in the model (including the intercept).
\item \code{k}: number of studies included in the analysis.
\item \code{k.all}: total number of studies (before exclusion).
\item \code{kdz},\code{ksz}: number of double-zero and single-zero studies.
\item \code{k1sz}, \code{k2sz}: number of single-zero studies where the zero is in group 1 or group 2.
\item \code{ai}, \code{bi}, \code{ci}, \code{di}: original entries of the 2x2 tables for all studies.
\item \code{ni}, \code{n1i}, \code{n2i}: original total and group sample sizes.
\item ...
}
}
\description{
Function to conduct a meta-analysis of a rare events using a beta-binomial model.
See below for more details on this model and its application in meta-analyses of
rare events.
}
\details{
\loadmathjax{}
\subsection{Data input}{

The data input can be specified either through the arguments \code{ai},\code{bi},\code{ci},\code{di},\code{n1i}, and \code{n2i} (columns of the data frame \code{data}) or through the argument \code{x},
which takes an object that results from applying the \code{rareDescribe()} function to the data (i.e., the input for argument \code{x} must be an object of type \code{rareData}).
A \code{rareData} object can be produced from a data frame by applying the \code{rareDescribe()} function to it.
The \code{rareDescribe()} function pre-processes the data frame and stores the information required by the \code{rareBetabin()} function in a list.
See \code{?rareDescribe} for more details.
}

\subsection{Effect size measures}{

The function includes different versions of the beta-binomial model (see Kuss (2014), for a description).
The regression equation used in all these models can be expressed as

\mjseqn{g(\mu_j) = \alpha + \theta \cdot x_{ij}},

where \mjseqn{i = 1, ..., k } is the study index and \mjseqn{j = 1, 2} is the group index,
\mjseqn{x_{i1} = 1} and \mjseqn{x_{i2} = 0}.

All models assume that the number of events in each group follows a beta-binomial distribution
with mean \mjseqn{n_{ij}\mu_j} and intraclass correlation \mjseqn{\rho_j}.

For \code{measure = "logOR"}, \mjseqn{g(\cdot)} corresponds to the logit function, i.e., \mjseqn{g(\mu_j) = \log \left( \frac{\mu_j}{1-\mu_j} \right)}.

For \code{measure = "logRR"}, \mjseqn{g(\cdot)} correspond to the natural logarithm.

It is currently not possible to use \code{rareBetabin} with \code{measure = "RD"},
but this functionality may be implemented in future versions of this package.
}

\subsection{Group-specific intraclass correlation}{

Per default, it is assumed that \mjseqn{\rho_1 = \rho_2}, i.e. the intraclass correlations
are assumed to be equal for both groups. It is possible to fit the beta-binomial
with different intraclass correlations by setting \code{common_rho} to \code{FALSE}.
Internally, different intraclass correlations are modeled via the regression equation

\mjseqn{\log\left(\frac{\rho_j}{1-\rho_j}\right) = \zeta + \gamma \cdot x_{ij}},

where \mjseqn{x_{i1} = 1} and \mjseqn{x_{i2} = 0}. Estimates
for \mjseqn{\rho_1} and \mjseqn{\rho_2} are then obtained by back-transforming
the results to the original scale.
}
}
\examples{

# indtroducing the data
data(dat.nissen2007)
d <- dat.nissen2007

# estimating the log relative risk assuming common intraclass correlations
rareBetabin(ai=miRosiglitazone, ci=miControl, n1i=nRosiglitazone, n2i=nControl, data=d, measure="logRR", common_rho=TRUE)

# estimatint the log odds ratio assuming differing intraclass correlations
rareBetabin(ai=miRosiglitazone, ci=miControl, n1i=nRosiglitazone, n2i=nControl, data=d, measure="logOR", common_rho=FALSE)

# same analysis with pre-processed data
x <- rareDescribe(ai=miRosiglitazone, ci=miControl, n1i=nRosiglitazone, n2i=nControl, data=d)
rareBetabin(x, measure="logOR", common_rho=FALSE)

}
\references{
Kuss, O. (2014). Statistical methods for meta-analyses including information from studies
without any events-add nothing to nothing and succeed nevertheless. Statistics in
Medicine, 34 (7), 1097–1116. doi: 10.1002/sim.6383
}
