% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareDescribe.R
\name{rareDescribe}
\alias{rareDescribe}
\title{Descriptives for a meta-analysis of a rare event}
\usage{
rareDescribe(ai, bi, ci, di, n1i, n2i, data)
}
\arguments{
\item{ai}{Data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{Data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{Data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{Data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{Data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{Data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{Data frame.}
}
\value{
An object of the class rareData, which contains descriptives of the meta-analytic data.
}
\description{
Descriptives for a meta-analysis of a rare event
}
\examples{
data <- data.frame(
ai = c(0,3,2,0),
bi = c(20,18,15,19),
ci = c(1,4,0,0),
di = c(19,17,16,20)
)

rareDescribe(ai = ai, bi = bi, ci = ci, di = di, data = data)
}
