% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareDescribe.R
\name{rareDescribe}
\alias{rareDescribe}
\title{Descriptives for a meta-analysis of a rare event}
\usage{
rareDescribe(ai, bi, ci, di, n1i, n2i, data)
}
\arguments{
\item{ai}{data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{data frame.}
}
\value{
An object of the class \code{"rareData"}, which contains descriptives of the meta-analytic data.
}
\description{
Function to compute descriptives based on the 2x2 tables of the individual studies
for meta-analytic rare event data.
}
\examples{
data <- data.frame(
  ai = c(0, 3, 2, 0),
  bi = c(20, 18, 15, 19),
  ci = c(1, 4, 0, 0),
  di = c(19, 17, 16, 20)
)

desc <- rareDescribe(ai = ai, bi = bi, ci = ci, di = di, data = data)
print(desc)
}
