% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareGLMM.R
\name{rareGLMM}
\alias{rareGLMM}
\title{Conduct a meta-analysis using a generalized linear (mixed) model (GLMM)}
\usage{
rareGLMM(
  x,
  ai,
  bi,
  ci,
  di,
  n1i,
  n2i,
  data,
  measure,
  intercept = "fixed",
  slope = "random",
  conditional = FALSE,
  approx = FALSE,
  cor = FALSE,
  coding = 1/2,
  drop00 = FALSE,
  level = 95,
  test = "z",
  digits = 4,
  verbose = FALSE,
  control,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{ai}{data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{data frame.}

\item{measure}{character string specifying the effect size or outcome measure to be used
(either \code{"logOR"} for the log odds ratio or \code{"logRR"} for the log relative risk). See below for more details.}

\item{intercept}{character string specifying whether to fit a model with fixed, study-specific
intercepts (\code{"fixed"}) or a random intercept (\code{"random"}). See below for more details.}

\item{slope}{character string specifying whether to fit a model with a fixed slope (\code{"fixed"}) or a
random slope (\code{"random"}). A model with a fixed slope is a model with homogeneous effects, that is, a
fixed-effects model, while a model with a random slope is a model with heterogeneous effects, that is, a
random effects model. See below for more details.}

\item{conditional}{logical specifying whether to estimate a conditional generalized linear mixed model.
Default is \code{FALSE}.}

\item{approx}{logical specifying whether to use the approximate version of the conditional generalized linear
mixed model (i.e., the conditional binomial model). Only relevant when \code{conditional = TRUE} and \verb{measure = "logOR}.
Default is \code{FALSE}. See below for more details.}

\item{cor}{logical specifying whether random effects should be modeled as correlated or uncorrelated.
Default is \code{cor = FALSE}. This argument is only relevant if \code{intercept = "random"} and \code{slope = "random"}.
See below for more details.}

\item{coding}{numeric specifying the coding scheme used for the random effects structure. Values between 0
and 1 can be specified. Default is \code{coding = 1/2}. Given that \code{cor = FALSE}is specified, the default
option implies equal variances in the two groups. Values closer to 0 imply a larger variance in the
control group (group 2), while values closer to 1 imply a larger variance in the treatment group (group 1).
See below for more details.}

\item{drop00}{logical indicating whether double-zero studies (i.e., studies with no events or only events in both groups)
should be excluded prior to calculating the studies' effect sizes and sampling variances.}

\item{level}{numeric between 0 and 100 specifying the confidence interval level (the default is 95).}

\item{test}{character string specifying how test statistics and confidence intervals for the
fixed effects should be computed (currently, only \code{"z"}, for Wald-type tests is available).}

\item{digits}{integer specifying the number of decimal places to which the printed results
should be rounded (if unspecified, the default is 4).}

\item{verbose}{logical indicating whether output should be generated on the progress of model
fitting (the default is \code{FALSE}). Can also be an integer. Values > 1 generate more verbose output.}

\item{control}{optional list of control values for the iterative algorithms. If unspecified, default
values are defined inside the functions.}

\item{...}{additional arguments.}
}
\value{
an object of class "raremeta". The object is a list containing the following elements:
\itemize{
\item \code{model}: name of the model used for conducting the meta-analysis.
\item \code{beta}: estimated coefficients of the model.
\item \code{se}: standard errors of the  coefficients.
\item \code{zval}: test statistics of the coefficients.
\item \code{zval}: p-values corresponding to the test statistics.
\item \code{ci.lb}: lower bound of the confidence intervals for the coefficients.
\item \code{ci.ub}: upper bound of the confidence intervals for the coefficients.
\item \code{vb}: variance-covariance matrix of the estimated coefficients.
\item \code{tau2}: estimated amount of (residual) heterogeneity. Always \code{0} when \code{method = "FE"}.
\item \code{LRT.Chisq}: Test statistic of the likelihood ratio test testing for homogeneity.
\item \code{LRT.df}: Degrees of freedom of the likelihood ratio test testing for homogeneity.
\item \code{LRT.pval}: p-value of the likelihood ratio test testing for homogeneity.
\item \code{fit.stats}: a list with log-likelihood, deviance, AIC, BIC, and AICc values under
the unrestricted and restricted likelihood.
\item \code{p}: number of coefficients in the model (including the intercept).
\item \code{k}: number of studies included in the analysis.
\item \code{k.all}: total number of studies (before exclusion).
\item \code{kdz},\code{ksz}: number of double-zero and single-zero studies.
\item \code{k1sz}, \code{k2sz}: number of single-zero studies where the zero is in group 1 or group 2.
\item \code{ai}, \code{bi}, \code{ci}, \code{di}: original entries of the 2x2 tables for all studies.
\item \code{ni}, \code{n1i}, \code{n2i}: original total and group sample sizes.
\item ...
}
}
\description{
Function to conduct a meta-analysis of a rare event using a generalized linear
or generalized linear mixed model. See below for more details on these
models and their application in meta-analyses of rare events.
}
\details{
\loadmathjax{}
\subsection{Data input}{

The data input can be specified either through the arguments \code{ai},\code{bi},\code{ci},\code{di},\code{n1i}, and \code{n2i} (columns of the data frame \code{data})
or through the argument \code{x}, which takes an object that results from applying the \code{rareDescribe()} function to the data
(i.e., the input for argument \code{x} must be an object of type \code{rareData}).
A \code{rareData} object can be produced from a data frame by applying the \code{rareDescribe()} function to it.
The \code{rareDescribe()} function pre-processes the data frame and stores the information required by the \code{rareGLMM()} function in a list.
See \code{?rareDescribe} for more details.
}

\subsection{Effect size measures}{

The function includes different versions of the generalized linear (mixed) model. The regression equation
for the fixed-effects model (GLM) can be expressed as

\mjseqn{g(\pi_{ij}) = \alpha_i + \theta \cdot x_{ij}},

and the model equation for the random-effects model (GLMM) can be expressed as

\mjseqn{g(\pi_{ij}) = \alpha_i+ \theta \cdot x_{ij} + \epsilon_{i} \cdot z_{ij}},

where \mjseqn{i = 1, ..., k } is the study index and \mjseqn{j = 1, 2} is the group index,
\mjseqn{x_{i1} = 1}, \mjseqn{x_{i2} = 0}, and \mjseqn{z_{ij}} is defined by the \code{coding} argument.
Specifically, for \code{coding = z}, \mjseqn{z_{i1} = z} and \mjseqn{z_{i2} = z-1}. Default is \mjseqn{z = 1/2}.
\code{coding = 1} corresponds to Models 2 and 4 in Jackson et al. (2017), and \code{coding = 1/2} corresponds to Models 3 and 5 in
the same study.

For \code{measure = "logOR"}, a GL(M)M with a binomial within-study distribution is used for the numbers of
events in each group, with sample size \mjseqn{n_{ij}} and probability \mjseqn{\pi_{ij}}. For this model,
\mjseqn{g(\cdot)} corresponds to the logit function, i.e., \mjseqn{g(\pi_{ij}) = \log \left( \frac{\pi_{ij}}{1-\pi_{ij}} \right)}.
See Stijnen et al. (2011), for further details.

For \code{measure = "logRR"}, a GL(M)M with a Poisson within-study distribution is used for the numbers of
events in each group, with rate \mjseqn{n_{ij}\pi_{ij}}. For this model,
\mjseqn{g(\cdot)} corresponds to the natural logarithm. See Böhning et al. (2015), for
further details.

It is currently not possible to fit GL(M)Ms for \code{measure = "RD"}, but this functionality may be
enabled in future versions of the package.
}

\subsection{Baseline and effect heterogeneity}{

Baseline heterogeneity refers to the way the intercept is modeled. By specifying
\code{intercept = "fixed"}, the intercepts \mjseqn{\alpha_i} are modeled as fixed, study-specific
intercepts. By specifying \code{intercept = "random"}, it is assumed that \mjseqn{\alpha_i \sim N(\alpha, \sigma_{\alpha}^2)}.

Effect heterogeneity refers to the way the slope is modeled. By specifying
\code{slope = "fixed"}, a fixed slope is modeled, corresponding to a fixed-effects meta-analysis.
By specifying \verb{slope = "random}, it is assumed that \mjseqn{\epsilon_i \sim N(0, \tau^2)},
corresponding to a random-effects meta-analysis with between-study variance \mjseqn{\tau^2}.
}

\subsection{Correlated random effects}{

When both \code{intercept = "random"} and \code{slope = "random"} is specified, the argument \code{cor} can be
used to specify whether random-effects are assumed to be correlated. Per default, this is not
the case, i.e. \code{cor = FALSE}.
}
}
\examples{

# load a dataset
data(dat.nissen2007)
d <- dat.nissen2007

# GLMM for the log odds ratio with fixed intercept
rareGLMM(ai=miRosiglitazone, ci=miControl, n1i=nRosiglitazone, n2i=nControl, data=d, measure="logOR", intercept="fixed")

# same analysis with pre-processed data
x <- rareDescribe(ai = ai, bi = bi, ci = ci, di = di, data = data)
rareGLMM(x, measure="logOR", intercept="fixed")

}
\references{
Böhning, D., Mylona, K., & Kimber, A. (2015). Meta-analysis of clinical trials with rare
events. Biometrical Journal, 57 (4), 633–648. doi: 10.1002/bimj.201400184

Jackson, D., Law, M., Stijnen, T., Viechtbauer, W., & White, I. R. (2018). A comparison
of seven random-effects models for meta-analyses that estimate the summary odds
ratio. Statistics in Medicine, 37 (7), 1059–1085. doi: 10.1002/sim.7588

Stijnen, T., Hamza, T. H., & Özdemir, P. (2010). Random effects meta-analysis of event
outcome in the framework of the generalized linear mixed model with applications
in sparse data. Statistics in Medicine, 29 (29), 3046–3067. doi: 10.1002/sim.4040
}
