% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMH.R
\name{rareMH}
\alias{rareMH}
\title{Conduct a meta-analysis using Mantel-Haenszel type estimators}
\usage{
rareMH(
  x,
  ai,
  bi,
  ci,
  di,
  n1i,
  n2i,
  data,
  measure,
  level = 95,
  digits = 4,
  correct = FALSE,
  cc = "constant",
  ccval = 0.5,
  tccval,
  cccval,
  ccsum = 1,
  ccto = "only0",
  method = "FE"
)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{ai}{data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{data frame}

\item{measure}{character string specifying the effect size or outcome measure to be used
(either \code{"logOR"} for the log odds ratio, \code{"logRR"} for the log relative risk,
or \code{"RD"} for the risk difference).}

\item{level}{numeric between 0 and 100 specifying the confidence interval level (the default is 95).}

\item{digits}{integer specifying the number of decimal places to which the printed results
should be rounded (if unspecified, the default is 4).}

\item{correct}{logical specifying whether the data shall be continuity corrected before application of \code{rareMH()}. Default is \code{FALSE}}

\item{cc}{character string specifying the type of continuity correction to be used.
(either \code{"constant"}, \code{"tacc"} or \code{"empirical"}). Default is \code{"constant"}.}

\item{ccval}{scalar or numerical vector specifying the value of the continuity correction if
\code{cc = "constant"}. Must be a scalar or a vector of length equal to the number of studies.
Default is \code{ccval = 0.5}. If a scalar is specified, the value is added to all studies for
which the number of events is zero in at least one of the groups. This behavior can be changed
using the argument \code{ccto}. \code{ccval} is overwritten by tccval and cccval if both arguments are specified.}

\item{tccval}{scalar or numerical vector specifying the value of the continuity correction
applied to the observations from the treatment group (group 1) if \code{cc = "constant"}. Must be a scalar or a vector
of length equal to the number of studies. If \code{cc = "constant"} and \code{tccval} is not specified, \code{tccval} is
set to the value of \code{ccval} internally.}

\item{cccval}{scalar or numerical vector specifying the value of the continuity correction
applied to the observations from the control group (group 2) if \code{cc = "constant"}. Must be a scalar or a vector
of length equal to the number of studies. If \code{cc = "constant"} and \code{cccval} is not specified, \code{cccval} is
set to the value of \code{ccval} internally.}

\item{ccsum}{numeric value specifying the value of the sum of the continuity correction applied to the
observations from the treatment group and the continuity correction applied to the observations from
the control group. Default is \code{ccsum = 1}. Currently, setting this argument to a different number only has
an effect when \code{cc = "tacc"} or \code{cc = "empirical"}.}

\item{ccto}{character string indicating to which studies the continuity correction should
be applied. Either \code{"only0"}, for which the continuity correction is applied to all studies
for which the number of events is zero in at least one of the groups, \code{"all"}, for which the
continuity correction is applied to all studies, or \code{"if0all"}, for which the continuity
correction is applied to all studies if any of the individual studies has zero events in at
least one of the groups.}

\item{method}{character string specifying whether a fixed- or a random-effects model should be fitted when
calculating the empirical and treatment-arm continuity correction. See ´?rareCC()´ for more details.
A fixed-effects model is fitted when using \code{method = "FE"} . A random-effects model is fitted
by setting \code{method} equal to one of the following: \code{"DL"}, \code{"HE"}, \code{"SJ"}, \code{"ML"}, \code{"REML"}, \code{"EB"}, \code{"HS"},
\code{"PM"}, \code{"IPM"}, \code{"GENQ"}, \code{"PMM"} or \code{"GENQM"}.}

\item{drop00}{logical indicating whether double-zero studies (i.e., studies with no events or only events in both groups)
should be excluded prior to calculating the studies' effect sizes and sampling variances.}
}
\value{
an object of class "raremeta".
The object is a list containing the following elements:
\itemize{
\item \code{beta}, \code{b}: estimated effect size.
\item \code{se}: estimated standard error of the  estimator.
\item \code{zval}: test statistics of the coefficients.
\item \code{pval}: p-values corresponding to the test statistics.
\item \code{ci.lb}: lower bound of the confidence intervals for the coefficients.
\item \code{ci.ub}: upper bound of the confidence intervals for the coefficients.
\item \code{k}: number of studies included in the analysis.
\item \code{kdz},\code{ksz}: number of double-zero and single-zero studies.
\item \code{k1sz}, \code{k2sz}: number of single-zero studies where the zero is in group 1 or group 2.
\item \code{ai}, \code{bi}, \code{ci}, \code{di}: original entries of the 2x2 tables for all studies.
\item \code{ni}, \code{n1i}, \code{n2i}: original total and group sample sizes.
\item ...
}
}
\description{
Function to conduct a meta-analysis.
Effect sizes are the log odds ratio, log relative risk and
risk difference estimated from event counts in form of 2x2 contingency
tables using Mantel-Haenszel type estimators.
}
\section{Details}{
\subsection{Data input}{

The data input can be specified either through the arguments \code{ai},\code{bi},\code{ci},\code{di},\code{n1i}, and \code{n2i} (columns of the data frame \code{data})
or through the argument \code{x}, which takes an object that results from applying the \code{rareDescribe()} function to the data
(i.e., the input for argument \code{x} must be an object of type \code{rareData}).
A \code{rareData} object can be produced from a data frame by applying the \code{rareDescribe()} function to it.
The \code{rareDescribe()} function pre-processes the data frame and stores the information required by the \code{rareMH()} function in a list.
See \code{?rareDescribe} for more details.
}

\subsection{Effect size measures}{

The function includes meta-analytic methods for log odds ratios, log risk ratios, and risk differences.
The effect size measure can be specified using the \code{measure} argument. The respective effect size,
along with an estimate of its sampling variance, is then calculated for each study based on the
entries of the study's 2x2 table:\tabular{lrr}{
    \tab event \tab no event \cr
   group1 (treatment) \tab ai \tab bi \cr
   group2 (control) \tab ci \tab di \cr
}

}

\subsection{Mantel-Haenszel type estimators}{

A prominent class of estimators for the common effect size under the the equal-effects model
was proposed by Mantel and Haenszel in 1959.
Mantel-Haenszel type estimators are particularly useful in meta-analysis of rare
events.
As a quotient of sums rather than a sum of quotients,
the Mantel-Haenszel estimator of the overall effect is oftentimes defined even though the standard
estimator for the corresponding effect size for some individual studies may not be defined.
The variance estimation is due to Greenland et al. (1985) and
Robins et al. (1986).
For a summary of the implemented estimators see e.g. sections 9.2-9.4 of Jewell
(2013).
}
}

\examples{
# load a data set
data(dat.nissen2007)
d <- dat.nissen2007

# meta analysis of the log odds ratio using the Mantel-Haenszel method
rareMH(ai=miRosiglitazone, ci=miControl, n1i=nRosiglitazone, n2i=nControl, data=d, measure="logOR")

# same analysis with pre-processed data
x <- rareDescribe(ai=miRosiglitazone, ci=miControl, n1i=nRosiglitazone, n2i=nControl, data=d)
rareMH(x, measure="logOR")

}
\references{
Mantel, N., & Haenszel, W. (1959).
Statistical aspects of the analysis of data from retrospective studies of disease.
Journal of the national cancer institute, 22(4), 719-748.

Greenland, S., & Robins, J. M. (1985).
Estimation of a common effect parameter from sparse follow-up data.
Biometrics, 55-68.

Robins, J., Breslow, N., & Greenland, S. (1986).
Estimators of the Mantel-Haenszel variance consistent in both sparse data
and large-strata limiting models.
Biometrics, 311-323.

Jewell, N. P. (2003). Statistics for epidemiology. Chapman and Hall/CRC.
}
