% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareMH.R
\name{rareMH}
\alias{rareMH}
\title{Conduct a meta-analysis using Mantel-Haenszel type estimators}
\usage{
rareMH(x, ai, bi, ci, di, n1i, n2i, data, measure, level = 95, digits = 4)
}
\arguments{
\item{x}{an object of class \code{"rareData"}.}

\item{ai}{data frame column to specify the number of events in group 1 (i.e., the treatment group).}

\item{bi}{data frame column to specify the number of non-events in group 1 (i.e., the treatment group).}

\item{ci}{data frame column to specify the number of events in group 2 (i.e., the control group).}

\item{di}{data frame column to specify number of non-events in group 2 (i.e., the control group).}

\item{n1i}{data frame column to specify the sample sizes in group 1 (i.e., the treatment group).}

\item{n2i}{data frame column to specify the sample sizes in group 2 (i.e., the control group).}

\item{data}{data frame}

\item{measure}{character string specifying the effect size or outcome measure to be used
(either \code{"logOR"} for the log odds ratio, \code{"logRR"} for the log relative risk,
or \code{"RD"} for the risk difference).}

\item{level}{numeric between 0 and 100 specifying the confidence interval level (the default is 95).}

\item{digits}{integer specifying the number of decimal places to which the printed results
should be rounded (if unspecified, the default is 4).}
}
\value{
an object of class "raremeta".
The object is a list containing the following elements:
\itemize{
\item \code{beta}, \code{b}: estimated effect size.
\item \code{se}: estimated standard error of the  estimator.
\item \code{zval}: test statistics of the coefficients.
\item \code{pval}: p-values corresponding to the test statistics.
\item \code{ci.lb}: lower bound of the confidence intervals for the coefficients.
\item \code{ci.ub}: upper bound of the confidence intervals for the coefficients.
\item \code{k}: number of studies included in the analysis.
\item \code{kdz},\code{ksz}: number of double-zero and single-zero studies.
\item \code{k1sz}, \code{k2sz}: number of single-zero studies where the zero is in group 1 or group 2.
\item \code{ai}, \code{bi}, \code{ci}, \code{di}: original entries of the 2x2 tables for all studies.
\item \code{ni}, \code{n1i}, \code{n2i}: original total and group sample sizes.
\item ...
}
}
\description{
Function to conduct a meta-analysis.
Effect sizes are the log odds ratio, log relative risk and
risk difference estimated from event counts in form of 2x2 contingency
tables using Mantel-Haenszel type estimators.
}
\section{Details}{
\subsection{Data input}{

The data input can be specified either through the arguments \code{ai},\code{bi},\code{ci},\code{di},\code{n1i}, and \code{n2i} (columns of the data frame \code{data})
or through the argument \code{x}, which takes an object that results from applying the \code{rareDescribe()} function to the data
(i.e., the input for argument \code{x} must be an object of type \code{rareData}).
A \code{rareData} object can be produced from a data frame by applying the \code{rareDescribe()} function to it.
The \code{rareDescribe()} function pre-processes the data frame and stores the information required by the \code{rareMH()} function in a list.
See \code{?rareDescribe} for more details.
}

\subsection{Effect size measures}{

The function includes meta-analytic methods for log odds ratios, log risk ratios, and risk differences.
The effect size measure can be specified using the \code{measure} argument. The respective effect size,
along with an estimate of its sampling variance, is then calculated for each study based on the
entries of the study's 2x2 table:\tabular{lrr}{
    \tab event \tab no event \cr
   group1 (treatment) \tab ai \tab bi \cr
   group2 (control) \tab ci \tab di \cr
}

}

\subsection{Mantel-Haenszel type estimators}{

A prominent class of estimators for the common effect size under the the equal-effects model
was proposed by Mantel and Haenszel in 1959.
Mantel-Haenszel type estimators are particularly useful in meta-analysis of rare
events.
As a quotient of sums rather than a sum of quotients,
the Mantel-Haenszel estimator of the overall effect is oftentimes defined even though the standard
estimator for the corresponding effect size for some individual studies may not be defined.
The variance estimation is due to Greenland et al. (1985) and
Robins et al. (1986).
For a summary of the implemented estimators see e.g. sections 9.2-9.4 of Jewell
(2013).
}
}

\examples{
# load a data set
data(dat.nissen2007)
d <- dat.nissen2007

# meta analysis of the log odds ratio using the Mantel-Haenszel method
rareMH(ai=miRosiglitazone, ci=miControl, n1i=nRosiglitazone, n2i=nControl, data=d, measure="logOR")

# same analysis with pre-processed data
x <- rareDescribe(ai=miRosiglitazone, ci=miControl, n1i=nRosiglitazone, n2i=nControl, data=d)
rareMH(x, measure="logOR")

}
\references{
Mantel, N., & Haenszel, W. (1959).
Statistical aspects of the analysis of data from retrospective studies of disease.
Journal of the national cancer institute, 22(4), 719-748.

Greenland, S., & Robins, J. M. (1985).
Estimation of a common effect parameter from sparse follow-up data.
Biometrics, 55-68.

Robins, J., Breslow, N., & Greenland, S. (1986).
Estimators of the Mantel-Haenszel variance consistent in both sparse data
and large-strata limiting models.
Biometrics, 311-323.

Jewell, N. P. (2003). Statistics for epidemiology. Chapman and Hall/CRC.
}
